requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-bitcoinrpc>=1.0
aiofiles>=23.2.1
python-dotenv>=1.0.0
slowapi>=0.1.9
